+++ pkg/utils/README.md
@@ -1,0 +1,2 @@
+This is general utils for the plugin... this could be things we may want to move in to the SDK
+but need to be local until that version is up-to-date and/or the function is ready to share
